requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pydantic>=2.6.0
python-dotenv>=1.0.1
groq==0.4.2
